requirements.txt
pygame==2.6.1
PyOpenGL==3.1.9